@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.2.6'
+gem 'rails', '5.0.0'
 gem 'pg'
 gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
@@ -8,11 +8,9 @@ gem 'jquery-rails'
 gem 'jbuilder', '~> 2.0'
 gem 'sdoc', '~> 0.4.0', group: :doc
 gem 'devise'
-gem 'rails_admin'
 gem 'paperclip'
 gem 'friendly_id'
 gem 'acts_as_list'
-gem 'rails_admin_duration'
 gem 'redcarpet'
 
 group :development, :test do
@@ -20,6 +18,10 @@ group :development, :test do
   gem 'rspec-rails', '~> 3.4'
   gem 'factory_girl_rails'
   gem 'capybara'
+end
+
+group :test do
+  gem 'rails-controller-testing'
 end
 
 group :development do
